refactor: split Constants.jsx, relocate data files, and introduce SettingsContext
Phase 1 componentization (continued from bracket and dark mode extractions):

Split Constants.jsx (7,537 lines) into focused data modules:
- data/eventNames/frcEvents.js — FRC event name mappings (~990 lines)
- data/eventNames/ftcEvents.js — FTC event name mappings (~4,751 lines)
- data/eventNames/index.js — barrel re-export
- data/ftcRegions.js — FTC regions array (~454 lines)
- data/hallOfFame.js — Hall of Fame, special awards, alliance short names (~861 lines)
- data/matchClasses.js — bracket topology definitions (~408 lines)
- data/appConfig.js — refresh rate, community update template, playoff menu, etc. (~75 lines)
- Deleted the original Constants.jsx monolith

Relocated 4 pure-data files from components/ to data/:
- components/AppUpdates.jsx -> data/appUpdates.js
- components/Fouls.jsx -> data/fouls.js
- components/FoulsFTC.jsx -> data/foulsFTC.js
- components/TimeZones.jsx -> data/timeZones.js

Introduced SettingsContext for display/UX preferences:
- Created contexts/SettingsContext.jsx with 28 usePersistentState settings
- Uses useMemo to prevent unnecessary consumer re-renders
- Wired SettingsProvider into index.jsx provider stack
- App.jsx now consumes settings via useSettings() instead of local state
- SetupPage.jsx consumes settings via useSettings() instead of props
- Removed 56 settings-related props from SetupPage rendering (139 -> 83 props)

Updated all consumer imports across 12 files to reference new data/ paths.

Build passes with zero warnings. Bundle size reduced by 287 bytes.
