+++ src/data/appConfig.js
@@ -1,0 +1,75 @@
+export const communityUpdateTemplate = {
+  nameShortLocal: "",
+  cityStateLocal: "",
+  topSponsorsLocal: "",
+  topSponsorLocal: "",
+  sponsorsLocal: "",
+  organizationLocal: "",
+  robotNameLocal: "",
+  awardsLocal: "",
+  teamMottoLocal: "",
+  teamNotesLocal: "",
+  teamYearsNoCompeteLocal: "",
+  showRobotName: "",
+  teamNotes: "",
+  sayNumber: "",
+  awardsTextLocal: "",
+  lastUpdate: "",
+  source: "",
+};
+
+export const originalAndSustaining = [
+  "20",
+  "45",
+  "126",
+  "148",
+  "151",
+  "157",
+  "190",
+  "191",
+  "250",
+];
+export const allianceSelectionOrderBase = [
+  { number: 1, round: 1 },
+  { number: 2, round: 1 },
+  { number: 3, round: 1 },
+  { number: 4, round: 1 },
+  { number: 5, round: 1 },
+  { number: 6, round: 1 },
+  { number: 7, round: 1 },
+  { number: 8, round: 1 },
+  { number: 8, round: 2 },
+  { number: 7, round: 2 },
+  { number: 6, round: 2 },
+  { number: 5, round: 2 },
+  { number: 4, round: 2 },
+  { number: 3, round: 2 },
+  { number: 2, round: 2 },
+  { number: 1, round: 2 },
+  { number: 1, round: 3 },
+  { number: 2, round: 3 },
+  { number: 3, round: 3 },
+  { number: 4, round: 3 },
+  { number: 5, round: 3 },
+  { number: 6, round: 3 },
+  { number: 7, round: 3 },
+  { number: 8, round: 3 },
+];
+
+export const allianceSelectionBaseRounds = {
+  round1: [1, 2, 3, 4, 5, 6, 7, 8],
+  round2: [8, 7, 6, 5, 4, 3, 2, 1],
+  round3: [1, 2, 3, 4, 5, 6, 7, 8],
+  round4: [8, 7, 6, 5, 4, 3, 2, 1],
+};
+
+export const refreshRate = 15; //Refresh rate in seconds for auto update
+
+export const playoffOverrideMenu = [
+  { value: 8, label: 8 },
+  { value: 7, label: 7 },
+  { value: 6, label: 6 },
+  { value: 5, label: 5 },
+  { value: 4, label: 4 },
+];
+
