+++ src/data/eventNames/index.js
@@ -1,0 +1,2 @@
+export { eventNames } from "./frcEvents";
+export { FTCEventNames } from "./ftcEvents";
