+++ src/data/ftcRegions.js
@@ -1,0 +1,454 @@
+export const ftcRegions = [
+  {
+    regionCode: "ONADOD",
+    description: "Department of Defense Education Activity",
+  },
+  {
+    regionCode: "USAK",
+    description: "Alaska",
+  },
+  {
+    regionCode: "USAL",
+    description: "Alabama",
+  },
+  {
+    regionCode: "AU",
+    description: "Australia",
+  },
+  {
+    regionCode: "USAZ",
+    description: "Arizona",
+  },
+  {
+    regionCode: "BR",
+    description: "Brazil",
+  },
+  {
+    regionCode: "CAAB",
+    description: "Alberta - Canada",
+  },
+  {
+    regionCode: "USCHS",
+    description: "Chesapeake",
+  },
+  {
+    regionCode: "CY",
+    description: "Cyprus",
+  },
+  {
+    regionCode: "USMI",
+    description: "FIRST in Michigan",
+  },
+  {
+    regionCode: "USFL",
+    description: "Florida",
+  },
+  {
+    regionCode: "USHI",
+    description: "Hawaii",
+  },
+  {
+    regionCode: "USIA",
+    description: "Iowa",
+  },
+  {
+    regionCode: "USID",
+    description: "Idaho",
+  },
+  {
+    regionCode: "USIL",
+    description: "Illinois",
+  },
+  {
+    regionCode: "USIN",
+    description: "Indiana",
+  },
+  {
+    regionCode: "IN",
+    description: "India",
+  },
+  {
+    regionCode: "IL",
+    description: "Israel",
+  },
+  {
+    regionCode: "IT",
+    description: "Italy",
+  },
+  {
+    regionCode: "JM",
+    description: "Jamaica",
+  },
+  {
+    regionCode: "USMA",
+    description: "Massachusetts",
+  },
+  {
+    regionCode: "MA",
+    description: "Morocco",
+  },
+  {
+    regionCode: "USMN",
+    description: "Minnesota",
+  },
+  {
+    regionCode: "USCANO",
+    description: "Northern California",
+  },
+  {
+    regionCode: "USNE",
+    description: "Nebraska",
+  },
+  {
+    regionCode: "NG",
+    description: "Nigeria",
+  },
+  {
+    regionCode: "USNYNY",
+    description: "New York - NYC",
+  },
+  {
+    regionCode: "USNYEX",
+    description: "New York - Excelsior",
+  },
+  {
+    regionCode: "USNYLI",
+    description: "New York - Long Island",
+  },
+  {
+    regionCode: "USOH",
+    description: "Ohio",
+  },
+  {
+    regionCode: "USOK",
+    description: "Oklahoma",
+  },
+  {
+    regionCode: "CAON",
+    description: "Ontario",
+  },
+  {
+    regionCode: "USOR",
+    description: "Oregon",
+  },
+  {
+    regionCode: "USPA",
+    description: "Pennsylvania",
+  },
+  {
+    regionCode: "QA",
+    description: "Qatar",
+  },
+  {
+    regionCode: "CAQC",
+    description: "Québec - Canada",
+  },
+  {
+    regionCode: "RO",
+    description: "Romania",
+  },
+  {
+    regionCode: "USSC",
+    description: "South Carolina",
+  },
+  {
+    regionCode: "CASK",
+    description: "Saskatchewan - Canada",
+  },
+  {
+    regionCode: "TH",
+    description: "Tailand",
+  },
+  {
+    regionCode: "USVA",
+    description: "Virginia",
+  },
+  {
+    regionCode: "USVT",
+    description: "Vermont",
+  },
+  {
+    regionCode: "USWA",
+    description: "Washington",
+  },
+  {
+    regionCode: "USWI",
+    description: "Wisconsin",
+  },
+  {
+    regionCode: "USWV",
+    description: "West Virginia",
+  },
+  {
+    regionCode: "USWY",
+    description: "Wyoming",
+  },
+  {
+    regionCode: "ZA",
+    description: "South Africa",
+  },
+  {
+    regionCode: "USAR",
+    description: "Arkansas",
+  },
+  {
+    regionCode: "ECB",
+    description: "Antigua and Barbuda",
+  },
+  {
+    regionCode: "CABC",
+    description: "British Columbia",
+  },
+  {
+    regionCode: "USCA",
+    description: "California",
+  },
+  {
+    regionCode: "USCALA",
+    description: "California - Los Angeles",
+  },
+  {
+    regionCode: "USCASD",
+    description: "California - San Diego",
+  },
+  {
+    regionCode: "CN",
+    description: "China",
+  },
+  {
+    regionCode: "HKG",
+    description: "China - Hong Kong",
+  },
+  {
+    regionCode: "CMPZ2",
+    description: "Championship",
+  },
+  {
+    regionCode: "USCO",
+    description: "Colorado",
+  },
+  {
+    regionCode: "USCT",
+    description: "Connecticut",
+  },
+  {
+    regionCode: "USDE",
+    description: "Delaware",
+  },
+  {
+    regionCode: "DE",
+    description: "Germany",
+  },
+  {
+    regionCode: "FR",
+    description: "France",
+  },
+  {
+    regionCode: "USGA",
+    description: "Georgia - USA",
+  },
+  {
+    regionCode: "GB",
+    description: "United Kingdom",
+  },
+  {
+    regionCode: "JP",
+    description: "Japan",
+  },
+  {
+    regionCode: "KAZ",
+    description: "Kazakhstan",
+  },
+  {
+    regionCode: "USKS",
+    description: "Kansas",
+  },
+  {
+    regionCode: "USKY",
+    description: "Kentucky",
+  },
+  {
+    regionCode: "USLA",
+    description: "Louisiana",
+  },
+  {
+    regionCode: "LY",
+    description: "Libya",
+  },
+  {
+    regionCode: "HU",
+    description: "Hungary",
+  },
+  {
+    regionCode: "USMD",
+    description: "Maryland",
+  },
+  {
+    regionCode: "MD",
+    description: "Moldova",
+  },
+  {
+    regionCode: "MX",
+    description: "Mexico",
+  },
+  {
+    regionCode: "USMO",
+    description: "Missouri",
+  },
+  {
+    regionCode: "USMS",
+    description: "Mississippi",
+  },
+  {
+    regionCode: "USMT",
+    description: "Montana",
+  },
+  {
+    regionCode: "USNC",
+    description: "North Carolina",
+  },
+  {
+    regionCode: "USND",
+    description: "North Dakota",
+  },
+  {
+    regionCode: "USNH",
+    description: "New Hampshire",
+  },
+  {
+    regionCode: "USNJ",
+    description: "New Jersey",
+  },
+  {
+    regionCode: "NL",
+    description: "Netherlands",
+  },
+  {
+    regionCode: "USNM",
+    description: "New Mexico",
+  },
+  {
+    regionCode: "USNV",
+    description: "Nevada",
+  },
+  {
+    regionCode: "USNY",
+    description: "New York",
+  },
+  {
+    regionCode: "NZ",
+    description: "New Zealand",
+  },
+  {
+    regionCode: "PL",
+    description: "Poland",
+  },
+  {
+    regionCode: "USRI",
+    description: "Rhode Island",
+  },
+  {
+    regionCode: "SA",
+    description: "Saudi Arabia",
+  },
+  {
+    regionCode: "USSD",
+    description: "South Dakota",
+  },
+  {
+    regionCode: "KR",
+    description: "South Korea",
+  },
+  {
+    regionCode: "USTN",
+    description: "Tennessee",
+  },
+  {
+    regionCode: "TW",
+    description: "Taiwan",
+  },
+  {
+    regionCode: "USTX",
+    description: "Texas",
+  },
+  {
+    regionCode: "USTXHO",
+    description: "Texas - Houston",
+  },
+  {
+    regionCode: "USUT",
+    description: "Utah",
+  },
+  {
+    regionCode: "LT",
+    description: "Lithuania",
+  },
+  {
+    regionCode: "PY",
+    description: "Paraguay",
+  },
+  {
+    regionCode: "GR",
+    description: "Greece",
+  },
+  {
+    regionCode: "ZA",
+    description: "eSwatini",
+  },
+  {
+    regionCode: "ID",
+    description: "Indonesia",
+  },{
+    regionCode: "KZ",
+    description: "Kazakhstan",
+  },{
+    regionCode: "MY",
+    description: "Malaysia",
+  },{
+    regionCode: "USARL",
+    description: "Adventist Robotics League",
+  },{
+    regionCode: "USCALS",
+    description: "Los Angeles - California",
+  },{
+    regionCode: "USMOKS",
+    description: "Missouri-Kansas",
+  },{
+    regionCode: "USPR",
+    description: "Puerto Rico",
+  },{
+    regionCode: "VN",
+    description: "Vietnam",
+  },{
+    regionCode: "FPE",
+    description: "FTC Premier Events"
+  },{
+    regionCode: "NF",
+    description: "Newfoundland and Labrador - Canada"
+  },{
+    regionCode: "TR",
+    description: "Türkiye"
+  },{
+    regionCode: "USFCT",
+    description: "Four Corners Tribal Teams"
+  },{
+    regionCode: "EG",
+    description: "Egypt"
+  },{
+    regionCode: "USTXCE",
+    description: "Texas-Central"
+  },{
+    regionCode: "USTXNO",
+    description: "Texas-North"
+  },{
+    regionCode: "USTXSO",
+    description: "Texas-South"
+  },{
+    regionCode: "USTXWP",
+    description: "Texas-West & Panhandle"
+  },{
+    regionCode: "RU",
+    description: "Russian Federation"
+  }
+];
+
