+++ src/data/hallOfFame.js
@@ -1,0 +1,861 @@
+export const allianceShortNames = {
+  Archimedes: "Arch",
+  Tesla: "Tesla",
+  Darwin: "Darw",
+  Carson: "Cars",
+  Daly: "Daly",
+  Curie: "Curie",
+  Carver: "Carv",
+  Galileo: "Gal",
+  Hopper: "Hop",
+  Milstein: "Mil",
+  Johnson: "John",
+  Newton: "New",
+  Roebling: "Roeb",
+  Turing: "Tur",
+  "DTE Energy": "DTE",
+  Dow: "Dow",
+  "Dow ": "Dow",
+  Ford: "Ford",
+  Consumers: "Cons",
+};
+
+// Update this after Champs each season
+export const hallOfFame = [
+  {
+    Year: 2025,
+    Impact: 5985,
+    Challenge: "Reefscape Houston",
+    Winner1: 1323,
+    Winner2: 2910,
+    Winner3: 4272,
+    Winner4: 5026,
+    Winner5: null,
+  },
+  {
+    Year: 2024,
+    Impact: 2486,
+    Challenge: "Crescendo Houston",
+    Winner1: 1690,
+    Winner2: 4522,
+    Winner3: 9432,
+    Winner4: 321,
+    Winner5: null,
+  },
+  {
+    Year: 2023,
+    Impact: 321,
+    Challenge: "Charged Up Houston",
+    Winner1: 1323,
+    Winner2: 4414,
+    Winner3: 4096,
+    Winner4: 2609,
+    Winner5: null,
+  },
+  {
+    Year: 2022,
+    Chairmans: 1629,
+    Challenge: "Rapid React Houston",
+    Winner1: 1619,
+    Winner2: 254,
+    Winner3: 3175,
+    Winner4: 6672,
+    Winner5: null,
+  },
+  {
+    Year: 2021,
+    Chairmans: 503,
+    Challenge: "Infinite Recharge at Home",
+    Winner1: null,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2021,
+    Chairmans: 4613,
+    Challenge: "Infinite Recharge",
+    Winner1: null,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2019,
+    Chairmans: 1816,
+    Challenge: "Destination: Deep Space Detroit",
+    Winner1: 3707,
+    Winner2: 217,
+    Winner3: 4481,
+    Winner4: 1218,
+    Winner5: null,
+  },
+  {
+    Year: 2019,
+    Chairmans: 1902,
+    Challenge: "Destination: Deep Space Houston",
+    Winner1: 973,
+    Winner2: 1323,
+    Winner3: 5026,
+    Winner4: 4201,
+    Winner5: null,
+  },
+  {
+    Year: 2018,
+    Chairmans: 1311,
+    Challenge: "PowerUP! Houston",
+    Winner1: 254,
+    Winner2: 148,
+    Winner3: 2976,
+    Winner4: 3075,
+    Winner5: null,
+  },
+  {
+    Year: 2018,
+    Chairmans: 2834,
+    Challenge: "PowerUP! Detroit",
+    Winner1: 2767,
+    Winner2: 27,
+    Winner3: 4027,
+    Winner4: 2708,
+    Winner5: null,
+  },
+  {
+    Year: 2017,
+    Chairmans: 3132,
+    Challenge: "STEAMworks Houston",
+    Winner1: 973,
+    Winner2: 1011,
+    Winner3: 2928,
+    Winner4: 5499,
+    Winner5: null,
+  },
+  {
+    Year: 2017,
+    Chairmans: 2614,
+    Challenge: "STEAMworks St Louis",
+    Winner1: 2767,
+    Winner2: 254,
+    Winner3: 862,
+    Winner4: 1676,
+    Winner5: null,
+  },
+  {
+    Year: 2017,
+    Chairmans: null,
+    Challenge: "Festival of Champions",
+    Winner1: 2767,
+    Winner2: 254,
+    Winner3: 862,
+    Winner4: 1676,
+    Winner5: null,
+  },
+  {
+    Year: 2016,
+    Chairmans: 987,
+    Challenge: "FIRST Stronghold",
+    Winner1: 330,
+    Winner2: 2481,
+    Winner3: 120,
+    Winner4: 1086,
+    Winner5: null,
+  },
+  {
+    Year: 2015,
+    Chairmans: 597,
+    Challenge: "Recycle Rush",
+    Winner1: 118,
+    Winner2: 1671,
+    Winner3: 1678,
+    Winner4: 5012,
+    Winner5: null,
+  },
+  {
+    Year: 2014,
+    Chairmans: 27,
+    Challenge: "Aerial Assist",
+    Winner1: 254,
+    Winner2: 469,
+    Winner3: 2848,
+    Winner4: 74,
+    Winner5: null,
+  },
+  {
+    Year: 2013,
+    Chairmans: 1538,
+    Challenge: "ULTIMATE ASCENT",
+    Winner1: 1241,
+    Winner2: 1477,
+    Winner3: 610,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2012,
+    Chairmans: 1114,
+    Challenge: "Rebound Rumble",
+    Winner1: 180,
+    Winner2: 16,
+    Winner3: 25,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2011,
+    Chairmans: 359,
+    Challenge: "LogoMotion",
+    Winner1: 254,
+    Winner2: 111,
+    Winner3: 973,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2010,
+    Chairmans: 341,
+    Challenge: "Breakaway!",
+    Winner1: 294,
+    Winner2: 67,
+    Winner3: 177,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2009,
+    Chairmans: 236,
+    Challenge: "Lunacy",
+    Winner1: 111,
+    Winner2: 67,
+    Winner3: 971,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2008,
+    Chairmans: 842,
+    Challenge: "FIRST Overdrive",
+    Winner1: 1114,
+    Winner2: 217,
+    Winner3: 148,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2007,
+    Chairmans: 365,
+    Challenge: "Rack-n-Roll",
+    Winner1: 177,
+    Winner2: 190,
+    Winner3: 987,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2006,
+    Chairmans: 111,
+    Challenge: "Aim High",
+    Winner1: 217,
+    Winner2: 296,
+    Winner3: 522,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2005,
+    Chairmans: 67,
+    Challenge: "Triple Play",
+    Winner1: 330,
+    Winner2: 67,
+    Winner3: 503,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2004,
+    Chairmans: 254,
+    Challenge: "FIRST Frenzy - Raising the Bar",
+    Winner1: 71,
+    Winner2: 435,
+    Winner3: 494,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2003,
+    Chairmans: 103,
+    Challenge: "Stack Attack",
+    Winner1: 65,
+    Winner2: 111,
+    Winner3: 469,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2002,
+    Chairmans: 175,
+    Challenge: "Zone Zeal",
+    Winner1: 66,
+    Winner2: 71,
+    Winner3: 173,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2001,
+    Chairmans: 22,
+    Challenge: "Diabolical Dynamics",
+    Winner1: 71,
+    Winner2: 125,
+    Winner3: 279,
+    Winner4: 294,
+    Winner5: 365,
+  },
+  {
+    Year: 2000,
+    Chairmans: 16,
+    Challenge: "Co-Opertition FIRST",
+    Winner1: 25,
+    Winner2: 232,
+    Winner3: 255,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1999,
+    Chairmans: 120,
+    Challenge: "Double Trouble",
+    Winner1: 1,
+    Winner2: 48,
+    Winner3: 176,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1998,
+    Chairmans: 23,
+    Challenge: "Ladder Logic",
+    Winner1: 45,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1997,
+    Chairmans: 47,
+    Challenge: "Toroid Terror",
+    Winner1: 71,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1996,
+    Chairmans: 144,
+    Challenge: "Hexagon Havoc",
+    Winner1: 73,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1995,
+    Chairmans: 151,
+    Challenge: "Ramp-n-Roll",
+    Winner1: 100,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1994,
+    Chairmans: 191,
+    Challenge: "Tower Power ",
+    Winner1: 144,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1993,
+    Chairmans: 7,
+    Challenge: "Rug Rage",
+    Winner1: 148,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 1992,
+    Chairmans: 191,
+    Challenge: "Maize Craize",
+    Winner1: 126,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+];
+
+// Update this after Champs each season
+export const FTCHallOfFame = [
+  {
+    Year: 2025,
+    Inspire: 18139,
+    Challenge: "INTO THE DEEP",
+    Winner1: 19746,
+    Winner2: 11260,
+    Winner3: 20870,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2024,
+    Inspire: 12791,
+    Challenge: "CENTERSTAGE",
+    Winner1: 19066,
+    Winner2: 11212,
+    Winner3: 18763,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2023,
+    Inspire: 18438,
+    Challenge: "POWER PLAY",
+    Winner1: 11260,
+    Winner2: 20870,
+    Winner3: 19746,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2022,
+    Inspire: 8565,
+    Challenge: "FREIGHT FRENZY",
+    Winner1: 17713,
+    Winner2: 11260,
+    Winner3: 14725,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2021,
+    Inspire: null,
+    Challenge: "Ultimate Goal Detroit",
+    Winner1: null,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2021,
+    Inspire: null,
+    Challenge: "Ultimate Goal Houston",
+    Winner1: null,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2020,
+    Inspire: null,
+    Challenge: "SKYSTONE Detroit",
+    Winner1: null,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2020,
+    Inspire: null,
+    Challenge: "SKYSTONE Houston",
+    Winner1: null,
+    Winner2: null,
+    Winner3: null,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2019,
+    Inspire: null,
+    Challenge: "Rover Ruckus Houston",
+    Winner1: 3101,
+    Winner2: 5064,
+    Winner3: 6547,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2019,
+    Inspire: null,
+    Challenge: "Rover Ruckus Detroit",
+    Winner1: 9971,
+    Winner2: 11115,
+    Winner3: 10091,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2018,
+    Inspire: null,
+    Challenge: "Relic Recovery Houston",
+    Winner1: 7750,
+    Winner2: 724,
+    Winner3: 12611,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2018,
+    Inspire: null,
+    Challenge: "Relic Recovery St Louis",
+    Winner1: 8644,
+    Winner2: 8393,
+    Winner3: 4029,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2017,
+    Inspire: null,
+    Challenge: "Festival of Champions",
+    Winner1: 724,
+    Winner2: 4216,
+    Winner3: 8651,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2017,
+    Inspire: null,
+    Challenge: "Velocity Vortex St. Louis",
+    Winner1: 8686,
+    Winner2: 6929,
+    Winner3: 5916,
+    Winner4: null,
+    Winner5: null,
+  },
+  {
+    Year: 2017,
+    Inspire: null,
+    Challenge: "Velocity Vortex Houston",
+    Winner1: 724,
+    Winner2: 4216,
+    Winner3: 8651,
+    Winner4: null,
+    Winner5: null,
+  },
+];
+
+export const specialAwards = [
+  {
+    Year: 2021,
+    awards: [
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Winner",
+        series: 1,
+        teamNumber: 5553,
+        schoolName: "Notre Dame de Bellegarde",
+        fullTeamName:
+          "Groupe NOËL/SDRA/Elior/Région Auvergne-Rhône-Alpes/Immobilière Notre Dame de Bellegarde/Lions Club/Crédit Mutuel/JEEDOM/IDRAC Business School/TraceParts/RS Components/Party Lite/Mairie de Collonges au Mont d'Or/MAHP/THALES/HEXIS/BBL GROUPE/ROTARY/Groupe KARDOL/CIC/ELCIA/SOLMUREX&Notre Dame de Bellegarde",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 5553,
+        schoolName: "Notre Dame de Bellegarde",
+        fullTeamName:
+          "Groupe NOËL/SDRA/Elior/Région Auvergne-Rhône-Alpes/Immobilière Notre Dame de Bellegarde/Lions Club/Crédit Mutuel/JEEDOM/IDRAC Business School/TraceParts/RS Components/Party Lite/Mairie de Collonges au Mont d'Or/MAHP/THALES/HEXIS/BBL GROUPE/ROTARY/Groupe KARDOL/CIC/ELCIA/SOLMUREX&Notre Dame de Bellegarde",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 8,
+        schoolName: "Palo Alto High School",
+        fullTeamName:
+          "Apple/Bayer Fund/KLA Foundation/Arm Inc./TE Connectivity/BAE Systems/Intuitive Foundation/Markforged/Robby Technologies/Notion/Kirk’s Steakburgers/Emotiv/Solidworks/Asian Box/D&K Precision Sheet Metal/Dropbox&Palo Alto High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 172,
+        schoolName: "Falmouth High School & Gorham High School",
+        fullTeamName:
+          "Lanco Integrated / Chinburg Management / Falmouth High School / Adobe / Texas Instruments / Robotics Institute of Maine / IDEXX Laboratories / GoFar & Falmouth High School & Gorham High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 900,
+        schoolName:
+          "North Carolina School of Science and Mathematics & Durham Public Schools & Local Home Schools",
+        fullTeamName:
+          "NCSSM Foundation/NVIDIA/Analog Devices/Google/VMware Foundation/SMT&North Carolina School of Science and Mathematics&Durham Public Schools&Local Home Schools",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 971,
+        schoolName: "Mountain View High School",
+        fullTeamName:
+          "NASA/NASA Ames Research Center/Google/Abbott Fund/Auris Health/DSS SolidWorks/The Linley Group/Lockheed-Martin/BAE Systems/Mountain View High School PTSA/Markforged Inc./Intuitive Surgical/West Coast Products/Automation Direct/Creative Manufacturing Solutions/Group Manufacturing Services/Plug and Play Datacenters&Mountain View High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 1374,
+        schoolName: "OAKVILLE TRAFALGAR HIGH SCHOOL",
+        fullTeamName:
+          "Halton District School Board - Technological Education (HDSB TE) / Canadian National Exhibition (CNE) / Hatch / Siemens Canada / Sheridan College (Sheridan FIRST Alumni) / Jobmaster Magnets / Tetra Society / Sinn Development &OAKVILLE TRAFALGAR HIGH SCHOOL",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 2554,
+        schoolName: "John P Stevens High School",
+        fullTeamName:
+          "Picatinny Arsenal / BAE Systems / Middlesex County College / Edison Public School District / Sekuler STEM Fund / The Shoppes at Colonial Village & John P Stevens High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 3616,
+        schoolName:
+          "Southside High School & Thibodaux STEM Magnet Acad & Ovey Comeaux High School & Lafayette High School",
+        fullTeamName:
+          "Comeaux Academic Enrichment Program/Lafayette Parish School System/ISA Lafayette/IEEE Lafayette/Pugh Family Foundation/CLECO/McGill Precision Waterjet LLC/Automated Production/TRINITY GRAPHX SIGNS & DESIGNS/CAD Solutions/Gordan Technologies LLC/Jack and Stephanie Godwin/Chitimacha Louisiana Open & Thibodaux STEM Magnet Acad & Southside High School & Ovey Comeaux High School & Lafayette High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 3792,
+        schoolName: "Family/Community",
+        fullTeamName:
+          "3M/Boeing/University of Missouri-Columbia, College of Engineering/Crawford's Inc.&Family/Community",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 4253,
+        schoolName: "Taipei American School",
+        fullTeamName: "Taipei American School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 4586,
+        schoolName: "Ironi B Modi'in",
+        fullTeamName:
+          "Apple/Google/Suzuki/ShaGal Marketing Solutions Ltd./Masaeiot/Dfus Natali/Ofer Falafel/Ledico/Bank Hapoalim/Pet-it/Ituran/The Association for the Elderly in Modi'in/Ministry of Education/Modi'in-Maccabim-Re'ut Municipality&Ironi B Modi'in",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 5406,
+        schoolName: "Bishop Ryan Catholic Secondary School",
+        fullTeamName:
+          "HWCDSB/John Deere Canada/Heritage Green Trust/Arcellor Mittal Dofasco/Jantz Canada&Bishop Ryan Catholic Secondary School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 5940,
+        schoolName: "Design Tech High School",
+        fullTeamName:
+          "Oracle/DDS Solidworks/TSSP/Auris Surgical Robotics Inc./Google/KC Tool&Design Tech High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 6418,
+        schoolName: "Family/Community",
+        fullTeamName: "Cruise Automation/Apple/BAE&Family/Community",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 6479,
+        schoolName: "Corona Del Sol High School",
+        fullTeamName:
+          "Nikola Motors/Dassault Systèmes SOLIDWORKS Corp./Industrial Metal Supply Co./Copper State Bolt & Nut Company/Professional Plastics/Wishes in Stitches Embroidery and Screen Printing/Corona del Sol Math Department/ABF Printing and Marketing/McCarthy Building Companies, Inc/Tempe Union High School District/Corona del Sol High School/Active Solutions, LLC/AZ Diamondbacks Foundation/Boeing/Port Plastics/Teletech Community Foundation/General Dynamics & Corona Del Sol High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 6652,
+        schoolName:
+          "Facultad de Ingeniería Mecánica y Eléctrica & Universidad Autonoma de Nuevo León",
+        fullTeamName:
+          "Softtek/Grainger S.A. de C.V./SATI/Carto Monterrey S.A. de C.V./Atlas Management S.A. de C.V./Jetro Systemic S.A. de C.V.&Universidad Autonoma de Nuevo León&Facultad de Ingeniería Mecánica y Eléctrica",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 7480,
+        schoolName: "Family/Community",
+        fullTeamName:
+          "2019 Argosy Foundation FRC Rookie Grant/Novelis Inc./FIRST Canada & Family/Community",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 7504,
+        schoolName: "Brewster High School",
+        fullTeamName:
+          "Collins Aerospace/NASA/Danbury Walmart&Brewster High School",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 7563,
+        schoolName: "SESI",
+        fullTeamName: "SESI",
+        person: null,
+      },
+      {
+        awardId: null,
+        teamId: null,
+        eventId: null,
+        eventDivisionId: null,
+        eventCode: "INNCE",
+        name: "Innovation Challenge Finalist",
+        series: 1,
+        teamNumber: 8621,
+        schoolName: "Family/Community",
+        fullTeamName: "Family/Community",
+        person: null,
+      },
+    ],
+  },
+];
+
