+++ src/data/matchClasses.js
@@ -1,0 +1,408 @@
+export const matchClassesBase = {
+  eightAlliance: [
+    {
+      matchNumber: 1,
+      description: "Match 1 (R1) (#1)",
+      winnerTo: { matchNumber: 7, station: "red" },
+      loserTo: { matchNumber: 5, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 2,
+      description: "Match 2 (R1) (#2)",
+      winnerTo: { matchNumber: 7, station: "blue" },
+      loserTo: { matchNumber: 5, station: "blue" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 3,
+      description: "Match 3 (R1) (#3)",
+      winnerTo: { matchNumber: 8, station: "red" },
+      loserTo: { matchNumber: 6, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 4,
+      description: "Match 4 (R1) (#4)",
+      winnerTo: { matchNumber: 8, station: "blue" },
+      loserTo: { matchNumber: 6, station: "blue" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 5,
+      description: "Match 5 (R2) (#5)",
+      winnerTo: { matchNumber: 10, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M1" },
+      blue: { class: "davidPriceWarning", from: "Lost M2" },
+    },
+    {
+      matchNumber: 6,
+      description: "Match 6 (R2) (#6)",
+      winnerTo: { matchNumber: 9, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M3" },
+      blue: { class: "davidPriceWarning", from: "Lost M4" },
+    },
+    {
+      matchNumber: 7,
+      description: "Match 7 (R2) (#7)",
+      winnerTo: { matchNumber: 11, station: "red" },
+      loserTo: { matchNumber: 9, station: "red" },
+      
+      red: { class: "success", from: "Won M1" },
+      blue: { class: "success", from: "Won M2" },
+    },
+    {
+      matchNumber: 8,
+      description: "Match 8 (R2) (#8)",
+      winnerTo: { matchNumber: 11, station: "blue" },
+      loserTo: { matchNumber: 10, station: "red" },
+      red: { class: "success", from: "Won M3" },
+      blue: { class: "success", from: "Won M4" },
+    },
+    {
+      matchNumber: 9,
+      description: "Match 9 (R3) (#9)",
+      winnerTo: { matchNumber: 12, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M7" },
+      blue: { class: "davidPriceWarning", from: "Won M6" },
+    },
+    {
+      matchNumber: 10,
+      description: "Match 10 (R3) (#10)",
+      winnerTo: { matchNumber: 12, station: "red" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M8" },
+      blue: { class: "davidPriceWarning", from: "Won M5" },
+    },
+    {
+      matchNumber: 11,
+      description: "Match 11 (R4) (#11)",
+      winnerTo: { matchNumber: 14, station: "red" },
+      loserTo: { matchNumber: 13, station: "red" },
+      red: { class: "success", from: "Won M7" },
+      blue: { class: "success", from: "Won M8" },
+    },
+    {
+      matchNumber: 12,
+      description: "Match 12 (R4) (#12)",
+      winnerTo: { matchNumber: 13, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Won M10" },
+      blue: { class: "davidPriceWarning", from: "Won M9" },
+    },
+    {
+      matchNumber: 13,
+      description: "Match 13 (R5) (#13)",
+      winnerTo: { matchNumber: 14, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M11" },
+      blue: { class: "davidPriceWarning", from: "Won M12" },
+    },
+    {
+      matchNumber: 14,
+      description: "Final 1 (#14)",
+      winnerTo: { matchNumber: 15, station: "blue" },
+      loserTo: { matchNumber: 15, station: "red" },
+    },
+    {
+      matchNumber: 15,
+      description: "Final 2 (#15)",
+      winnerTo: { matchNumber: 16, station: "blue" },
+      loserTo: { matchNumber: 16, station: "red" },
+    },
+    {
+      matchNumber: 16,
+      description: "Final 3 (#16)",
+      winnerTo: { matchNumber: 17, station: "blue" },
+      loserTo: { matchNumber: 17, station: "red" },
+    },
+    {
+      matchNumber: 17,
+      description: "Overtime 1 (#17)",
+      winnerTo: { matchNumber: 18, station: "blue" },
+      loserTo: { matchNumber: 18, station: "red" },
+    },
+    {
+      matchNumber: 18,
+      description: "Overtime 2 (#18)",
+      winnerTo: { matchNumber: 19, station: "blue" },
+      loserTo: { matchNumber: 19, station: "red" },
+    },
+    {
+      matchNumber: 19,
+      description: "Overtime 3 (#19)",
+      winnerTo: { matchNumber: null, station: null },
+      loserTo: { matchNumber: null, station: null },
+    },
+  ],
+  sixAlliance: [
+    {
+      matchNumber: 1,
+      description: "Match 1 (R1) (#1)",
+      winnerTo: { matchNumber: 3, station: "blue" },
+      loserTo: { matchNumber: 6, station: "blue" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 2,
+      description: "Match 2 (R1) (#2)",
+      winnerTo: { matchNumber: 4, station: "blue" },
+      loserTo: { matchNumber: 5, station: "blue" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 3,
+      description: "Match 3 (R1) (#3)",
+      winnerTo: { matchNumber: 7, station: "red" },
+      loserTo: { matchNumber: 5, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 4,
+      description: "Match 4 (R1) (#4)",
+      winnerTo: { matchNumber: 7, station: "blue" },
+      loserTo: { matchNumber: 6, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 5,
+      description: "Match 5 (R2) (#5)",
+      winnerTo: { matchNumber: 8, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M3" },
+      blue: { class: "davidPriceWarning", from: "Lost M2" },
+    },
+    {
+      matchNumber: 6,
+      description: "Match 6 (R2) (#6)",
+      winnerTo: { matchNumber: 8, station: "red" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M4" },
+      blue: { class: "davidPriceWarning", from: "Lost M1" },
+    },
+    {
+      matchNumber: 7,
+      description: "Match 7 (R2) (#7)",
+      winnerTo: { matchNumber: 10, station: "red" },
+      loserTo: { matchNumber: 9, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 8,
+      description: "Match 8 (R2) (#8)",
+      winnerTo: { matchNumber: 9, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M6" },
+      blue: { class: "davidPriceWarning", from: "Lost M5" },
+    },
+    {
+      matchNumber: 9,
+      description: "Match 9 (R3) (#9)",
+      winnerTo: { matchNumber: 10, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M7" },
+      blue: { class: "davidPriceWarning", from: "Lost M8" },
+    },
+    {
+      matchNumber: 10,
+      description: "Final 1 (#10)",
+      winnerTo: { matchNumber: 11, station: "blue" },
+      loserTo: { matchNumber: 11, station: "red" },
+    },
+    {
+      matchNumber: 11,
+      description: "Final 2 (#11)",
+      winnerTo: { matchNumber: 12, station: "blue" },
+      loserTo: { matchNumber: 12, station: "red" },
+    },
+    {
+      matchNumber: 12,
+      description: "Overtime 1 (#12)",
+      winnerTo: { matchNumber: 13, station: "blue" },
+      loserTo: { matchNumber: 13, station: "red" },
+    },
+    {
+      matchNumber: 13,
+      description: "Overtime 2 (#13)",
+      winnerTo: { matchNumber: 14, station: "blue" },
+      loserTo: { matchNumber: 14, station: "red" },
+    },
+    {
+      matchNumber: 14,
+      description: "Overtime 3 (#14)",
+      winnerTo: { matchNumber: null, station: null },
+      loserTo: { matchNumber: null, station: null },
+    },
+  ],
+  fourAlliance: [
+    {
+      matchNumber: 1,
+      description: "Match 1 (R1) (#1)",
+      winnerTo: { matchNumber: 3, station: "red" },
+      loserTo: { matchNumber: 4, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 2,
+      description: "Match 2 (R1) (#2)",
+      winnerTo: { matchNumber: 3, station: "blue" },
+      loserTo: { matchNumber: 4, station: "blue" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 3,
+      description: "Match 3 (R2) (#3)",
+      winnerTo: { matchNumber: 6, station: "red" },
+      loserTo: { matchNumber: 5, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 4,
+      description: "Match 4 (R2) (#4)",
+      winnerTo: { matchNumber: 5, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M1" },
+      blue: { class: "davidPriceWarning", from: "Lost M2" },
+    },
+    {
+      matchNumber: 5,
+      description: "Match 5 (R3) (#5)",
+      winnerTo: { matchNumber: 6, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M3" },
+      blue: { class: "davidPriceWarning", from: "Lost M4" },
+    },
+    {
+      matchNumber: 6,
+      description: "Final 1 (#6)",
+      winnerTo: { matchNumber: 7, station: "blue" },
+      loserTo: { matchNumber: 7, station: "red" },
+    },
+    {
+      matchNumber: 7,
+      description: "Final 2 (#7)",
+      winnerTo: { matchNumber: 8, station: "blue" },
+      loserTo: { matchNumber: 8, station: "red" },
+    },
+    {
+      matchNumber: 8,
+      description: "Final 3 (#8)",
+      winnerTo: { matchNumber: 9, station: "blue" },
+      loserTo: { matchNumber: 9, station: "red" },
+    },
+    {
+      matchNumber: 9,
+      description: "Overtime 1 (#9)",
+      winnerTo: { matchNumber: 10, station: "blue" },
+      loserTo: { matchNumber: 10, station: "red" },
+    },
+    {
+      matchNumber: 10,
+      description: "Overtime 2 (#10)",
+      winnerTo: { matchNumber: 11, station: "red" },
+      loserTo: { matchNumber: 11, station: "blue" },
+    },
+    {
+      matchNumber: 12,
+      description: "Overtime 3 (#11)",
+      winnerTo: null,
+      loserTo: null,
+    },
+  ],
+  
+  fourAllianceFTC: [
+    {
+      matchNumber: 1,
+      description: "Match 1 (R1) (#1)",
+      winnerTo: { matchNumber: 4, station: "red" },
+      loserTo: { matchNumber: 3, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 2,
+      description: "Match 2 (R1) (#2)",
+      winnerTo: { matchNumber: 4, station: "blue" },
+      loserTo: { matchNumber: 3, station: "blue" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 3,
+      description: "Match 3 (R2) (#3)",
+      winnerTo: { matchNumber: 5, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M1" },
+      blue: { class: "davidPriceWarning", from: "Lost M2" },
+      
+    },
+    {
+      matchNumber: 4,
+      description: "Match 4 (R2) (#4)",
+      winnerTo: { matchNumber: 6, station: "red" },
+      loserTo: { matchNumber: 5, station: "red" },
+      red: { class: "success", from: null },
+      blue: { class: "success", from: null },
+    },
+    {
+      matchNumber: 5,
+      description: "Match 5 (R3) (#5)",
+      winnerTo: { matchNumber: 6, station: "blue" },
+      loserTo: null,
+      red: { class: "davidPriceWarning", from: "Lost M3" },
+      blue: { class: "davidPriceWarning", from: "Lost M4" },
+    },
+    {
+      matchNumber: 6,
+      description: "Final 1 (#6)",
+      winnerTo: { matchNumber: 7, station: "blue" },
+      loserTo: { matchNumber: 7, station: "red" },
+    },
+    {
+      matchNumber: 7,
+      description: "Final 2 (#7)",
+      winnerTo: { matchNumber: 8, station: "blue" },
+      loserTo: { matchNumber: 8, station: "red" },
+    },
+    {
+      matchNumber: 8,
+      description: "Final 3 (#8)",
+      winnerTo: { matchNumber: 9, station: "blue" },
+      loserTo: { matchNumber: 9, station: "red" },
+    },
+    {
+      matchNumber: 9,
+      description: "Overtime 1 (#9)",
+      winnerTo: { matchNumber: 10, station: "blue" },
+      loserTo: { matchNumber: 10, station: "red" },
+    },
+    {
+      matchNumber: 10,
+      description: "Overtime 2 (#10)",
+      winnerTo: { matchNumber: 11, station: "red" },
+      loserTo: { matchNumber: 11, station: "blue" },
+    },
+    {
+      matchNumber: 12,
+      description: "Overtime 3 (#11)",
+      winnerTo: null,
+      loserTo: null,
+    },
+  ],
+
+};
